backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
rasterio>=1.4.0
numpy>=2.0.0
python-dotenv>=1.0.0
pyproj>=3.7.0
httpx>=0.28.0
google-genai>=1.0.0
Pillow>=11.0.0
earthengine-api>=1.0.0
geemap>=0.34.0
osmnx>=1.9.0

# Optional: For downloading real satellite/building data
# earthengine-api>=1.0.0
# geemap>=0.34.0
# osmnx>=1.9.0
